refactor: 전 계층 패키지 구조 정리 - 불필요한 패키지 제거 - 의도에 맞지 않거나 부적합한 패키지 제거 및 클래스 이동
