+++ application/src/main/java/dev/chan/api/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package dev.chan.common.exception;
+package dev.chan.api;
 
 public class GlobalExceptionHandler {
     // TODO: 글로벌 예외처리 작성 예정
