+++ application/src/main/java/dev/chan/api/file/dto/ErrorReason.java
@@ -1,4 +1,4 @@
-package dev.chan.api.file.response;
+package dev.chan.api.file.dto;
 
 public enum ErrorReason {
     REQUIRED, INVALID, NOT_FOUND
