+++ application/src/test/java/dev/chan/infrastructure/aws/PresignedUrlServiceTest.java
@@ -1,8 +1,8 @@
 package dev.chan.infrastructure.aws;
 
-import dev.chan.api.file.request.FileMetaDataDto;
-import dev.chan.application.command.PresignedUrlCommand;
-import dev.chan.application.command.PresignedUrlSpecification;
+import dev.chan.api.file.dto.FileMetaDataDto;
+import dev.chan.application.dto.PresignedUrlCommand;
+import dev.chan.application.dto.PresignedUrlSpecification;
 import dev.chan.application.file.PresignedUrlResponse;
 import dev.chan.application.file.PresignedUrlService;
 import dev.chan.application.file.S3KeyGenerator;
