+++ application/src/test/java/dev/chan/infrastructure/aws/S3PresignedUrlGeneratorTest.java
@@ -1,6 +1,6 @@
 package dev.chan.infrastructure.aws;
 
-import dev.chan.application.command.PresignedUrlSpecification;
+import dev.chan.application.dto.PresignedUrlSpecification;
 import dev.chan.application.file.PresignedUrlResponse;
 import dev.chan.application.file.S3PresignedUrlGenerator;
 import dev.chan.common.MimeType;
